requirements.txt
flask
flask-sock
Werkzeug
RPi.GPIO
picamera
gunicorn